Runtime: break circular dependency introduced by async datasource lookup
Importing getDataSourceInstanceSettings into DataSourceWithBackend closed a
loop, because settings.ts and dataSource.ts both import isExpressionReference
back from DataSourceWithBackend. Move ExpressionDatasourceRef and
isExpressionReference into a leaf module so the edge only runs one way, and
import the settings module directly rather than through the unstable barrel.
