+++ packages/grafana-runtime/src/utils/DataSourceWithBackend.ts
@@ -11,7 +11,6 @@ import {
   DataSourceApi,
   type DataSourceInstanceSettings,
   type DataSourceJsonData,
-  type DataSourceRef,
   getDataSourceRef,
   makeClassES5Compatible,
   parseLiveChannelAddress,
@@ -31,32 +30,14 @@ import {
   StreamingFrameAction,
   type StreamingFrameOptions,
 } from '../services';
+import { getDataSourceInstanceSettings } from '../services/dataSource/settings';
 
-import { getDataSourceInstanceSettings } from './../unstable';
+import { ExpressionDatasourceRef, isExpressionReference } from './expressionRef';
 import { publicDashboardQueryHandler } from './publicDashboardQueryHandler';
 import { isQueryServiceCompatible } from './qscheck';
 import { type BackendDataSourceResponse, toDataQueryResponse } from './queryResponse';
 import { UserStorage } from './userStorage';
 
-/**
- * @internal
- */
-export const ExpressionDatasourceRef = Object.freeze({
-  type: '__expr__',
-  uid: '__expr__',
-  name: 'Expression',
-});
-
-/**
- * @public
- */
-export function isExpressionReference(ref?: DataSourceRef | string | null): boolean {
-  if (!ref) {
-    return false;
-  }
-  const v = typeof ref === 'string' ? ref : ref.type;
-  return v === ExpressionDatasourceRef.type || v === ExpressionDatasourceRef.name || v === '-100'; // -100 was a legacy accident that should be removed
-}
 
 export class HealthCheckError extends Error {
   details: HealthCheckResultDetails;
